+++ CHANGELOG.md
@@ -8,9 +8,12 @@ with respect to its command line interface and HTTP interface
 
 
 ## [Unreleased](//github.com/opentable/sous/compare/0.5.93...HEAD)
+### Added
+* Server: state storage toggle behind a feature flag - servers can be
+  configured to use the database as the source of truth.
+
 ### Changed
 * Client: Slack add additional channels to send via config
-
 
 ## [0.5.93](//github.com/opentable/sous/compare/0.5.92...0.5.93)
 ### Changed